move deepspeed to `lib_integrations.deepspeed`
