+++ src/transformers/lib_integrations/__init__.py
@@ -12,3 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from .peft import PeftAdapterMixin
+from .deepspeed import deepspeed
