+++ src/transformers/lib_integrations/deepspeed/deepspeed.py
@@ -19,8 +19,8 @@ import importlib.util
 import weakref
 from functools import partialmethod
 
-from .dependency_versions_check import dep_version_check
-from .utils import is_accelerate_available, is_torch_available, logging
+from ...dependency_versions_check import dep_version_check
+from ...utils import is_accelerate_available, is_torch_available, logging
 
 
 if is_torch_available():
