+++ src/transformers/testing_utils.py
@@ -39,7 +39,7 @@ import requests
 
 from transformers import logging as transformers_logging
 
-from .deepspeed import is_deepspeed_available
+from .lib_integrations.deepspeed import is_deepspeed_available
 from .integrations import (
     is_clearml_available,
     is_fairscale_available,
