+++ src/transformers/trainer_seq2seq.py
@@ -20,7 +20,7 @@ import torch
 from torch import nn
 from torch.utils.data import Dataset
 
-from .deepspeed import is_deepspeed_zero3_enabled
+from .lib_integrations.deepspeed import is_deepspeed_zero3_enabled
 from .generation.configuration_utils import GenerationConfig
 from .trainer import Trainer
 from .utils import logging
